bugfix in scoring.py
when adding simulations we only simulate genes with a fitch score above a threshold, as only pairs of genes where both genes have at least a fitch score of k can produce simultaneous scores >= k. As we are considering pairs we need to adjust for this using the proportion of gene pairs where both have a fitch score >k.